Text component fast path
Truncated text an isolated component
